extract/requirements.txt
pandas>=2.0
requests>=2.28
openpyxl>=3.1
pyyaml>=6.0
